facebook-function/requirement.txt
pandas
google-cloud-storage
pandas-gbq
openpyxl
numpy